executor: create vecgroupcheck package to improve test
