+++ executor/BUILD.bazel
@@ -126,6 +126,7 @@ go_library(
         "//executor/internal/pdhelper",
         "//executor/internal/querywatch",
         "//executor/internal/util",
+        "//executor/internal/vecgroupchecker",
         "//executor/metrics",
         "//executor/mppcoordmanager",
         "//expression",
@@ -294,7 +295,6 @@ go_test(
         "charset_test.go",
         "chunk_size_control_test.go",
         "cluster_table_test.go",
-        "collation_test.go",
         "compact_table_test.go",
         "concurrent_map_test.go",
         "copr_cache_test.go",
