+++ executor/internal/vecgroupchecker/BUILD.bazel
@@ -1,0 +1,32 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "vecgroupchecker",
+    srcs = ["vecgroupchecker.go"],
+    importpath = "github.com/pingcap/tidb/executor/internal/vecgroupchecker",
+    visibility = ["//executor:__subpackages__"],
+    deps = [
+        "//expression",
+        "//sessionctx",
+        "//types",
+        "//util/chunk",
+        "//util/codec",
+    ],
+)
+
+go_test(
+    name = "vecgroupchecker_test",
+    timeout = "short",
+    srcs = ["vecgroupchecker_test.go"],
+    embed = [":vecgroupchecker"],
+    flaky = True,
+    shard_count = 3,
+    deps = [
+        "//expression",
+        "//parser/mysql",
+        "//types",
+        "//util/chunk",
+        "//util/mock",
+        "@com_github_stretchr_testify//require",
+    ],
+)
